requirements.txt
Flask
openai
python-dotenv
Flask-Cors
gunicorn